tests: mem_map: do not run on Intel Audio DSP SoCs [INTERNAL]
The mem_map test does not have cache manipulation so it does not
work on Intel Audio DSP SoCs. For now, exclude it until we have
sorted out all MMU issue, and this test is updated to include
cache manipulation.

The mem_map test does not have cache manipulation so it does not
work on Xtensa. For now, exclude it until we have sorted out
all MMU issue, and this test is updated to include cache
manipulation.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>

Signed-off-by: Daniel Leung <daniel.leung@intel.com>