tests: kernel/smp: Reduce iterations to pass on PTL sim
Some SMP tests (namely, `inc_concurrency` and `smp_switch_torture`) can
take a quite long time on the simulator. While simulation time is OK,
around 15 seconds, the real world elapsed time is some 30 times that,
and the test can take up almost 15 minutes. This patch changes the
'stress' factor of SMP tests to a tenth of original, which allows the
test to run on a reasonable amount of time.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>