tests: kernel/smp: Disable stress tests on LNL [INTERNAL]
While last reduction on iterations for the stress tests made them pass
locally, one of the stress tests (test_smp_switch_torture) is still
failing on CI environment (is machine busier?). This patch reduces even
further the iterations for the stress tests, to a factor of zero,
effectively disabling them.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>