ci: select building rimage based on parameters [INTERNAL]
Rimage module is only needed for audio devices.
The purpose of this task is to disable building
for platforms that do not need this module.
Added new entry in the targets.json file which
is of general use, whenever we want to add
an additional module to west, we can add or remove
it using a new line with the "west_project_filter": parameter.

Signed-off-by: Mateusz Junkier <mateusz.junkier@intel.com>