scripts: twister: Fix Windows pickle error
`target` and `args` for a `multiprocessing.Process` must be `pickle`d on OSes without `os.fork`,
e.g. Windows. If a method is passed there, it will be pickled along with its `__self__` argument,
which is its bounded owner. That owner is an instance of a class.
If that instance contains unpicklable fields, the pickling will fail.
Searching for such a problem is time-consuming,
as `pickle` does not specify which subfield caused the error.
It is safer to use fully static methods or functions if we need to pass any into a Process.

Signed-off-by: Lukasz Mrugala <lukaszx.mrugala@intel.com>